+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * server
   * naming scheme for context directories changed from app:tag to tag (current build is now _)
+  * build now creates a repository tag without app name
   
 
 ### 5.1.0 (pending)